backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Cerberus==1.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
distlib==0.3.7
Django==3.2.16
django-allauth==0.58.2
django-colorfield==0.11.0
django-extensions==3.2.3
django-filter==23.3
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
isort==5.13.0
oauthlib==3.2.2
pep517==0.13.1
Pillow==10.1.0
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requirementslib==3.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
urllib3==2.1.0
xlwt==1.3.0
yarg==0.1.9
